+++ receipt-analyzer-service/app/src/main/java/pl/madzierski/daniel/receipt/model/GetReceiptDetailsResponse.java
@@ -2,6 +2,7 @@ package pl.madzierski.daniel.receipt.model;
 
 import pl.madzierski.daniel.receipt.ReceiptResolverStrategyType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Collections;
@@ -98,10 +99,10 @@ public record GetReceiptDetailsResponse(
     public record ItemResponse(
         String id,
         String name,
-        Double amount,
-        Double unitPrice,
-        Double discount,
-        Double totalPrice,
+        BigDecimal amount,
+        BigDecimal unitPrice,
+        BigDecimal discount,
+        BigDecimal totalPrice,
         Integer position
     ) {
     }
@@ -111,7 +112,7 @@ public record GetReceiptDetailsResponse(
         ReceiptResolverStrategyType resolver,
         LocalDateTime createdDate,
         String brand,
-        Double totalPrice,
+        BigDecimal totalPrice,
         LocalDateTime payingDate,
         String address,
         Boolean isPreferredRevision,
@@ -125,7 +126,7 @@ public record GetReceiptDetailsResponse(
         ReceiptResolverStrategyType resolver,
         LocalDateTime createdDate,
         String brand,
-        Double totalPrice,
+        BigDecimal totalPrice,
         LocalDateTime payingDate,
         String address,
         Boolean isPreferredRevision,
