+++ receipt-analyzer-service/app/src/main/java/pl/madzierski/daniel/receipt/model/UpdateRevisionRequest.java
@@ -2,6 +2,7 @@ package pl.madzierski.daniel.receipt.model;
 
 import pl.madzierski.daniel.receipt.ReceiptResolverStrategyType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Set;
 
@@ -10,7 +11,7 @@ public record UpdateRevisionRequest(
     ReceiptResolverStrategyType resolver,
     LocalDateTime createdDate,
     String brand,
-    Double totalPrice,
+    BigDecimal totalPrice,
     LocalDateTime payingDate,
     String address,
     Boolean isPreferredRevision,
@@ -21,10 +22,10 @@ public record UpdateRevisionRequest(
     public record ItemRequest(
         String id,
         String name,
-        Double amount,
-        Double unitPrice,
-        Double discount,
-        Double totalPrice,
+        BigDecimal amount,
+        BigDecimal unitPrice,
+        BigDecimal discount,
+        BigDecimal totalPrice,
         Integer position
     ) {
     }
